Extract Controllers common methods to OreIdCallbacks concern
